# pkg/swagger.yaml
openapi: 3.0.3
info:
  title: Task service
  version: 1.0.0
  description: API for asynchronously processing tasks.
servers:
  - url: /
paths:
  /register:
    post:
      summary: Register user
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "201":
          description: User registered
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Username is already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /login:
    post:
      summary: Login user
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                type: object
                required: [token]
                properties:
                  token:
                    type: string
                    format: uuid
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /logout:
    post:
      summary: Logout user
      tags: [auth]
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Session terminated
        "401":
          $ref: "#/components/responses/Unauthorized"
  /task:
    post:
      summary: Create task
      tags: [tasks]
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: string
                    format: uuid
                required: [task_id]
  /status/{task_id}:
    get:
      summary: Get task status
      tags: [tasks]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TaskID"
      responses:
        "200":
          description: Task status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [in_progress, ready]
                required: [status]
        "404":
          $ref: "#/components/responses/NotFound"
  /result/{task_id}:
    get:
      summary: Get task result
      tags: [tasks]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TaskID"
      responses:
        "200":
          description: Task result
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
                required: [result]
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Task is not ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    TaskID:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    Credentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
  responses:
    BadRequest:
      description: Invalid request body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Task not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication is required or the token is invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"